+++ CHANGELOG.md
@@ -9,6 +9,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Demo cancer case gets loaded together with demo RD case in demo instance
 - Parse REVEL_score alongside REVEL_rankscore from csq field and display it on SNV variant page
 - Rank score results now show the ranking range
+- cDNA and protein changes displayed on institute causatives pages
 - Admin users can create missing beacon datasets from the institute's settings page
 ### Changed
 - Verify user before redirecting to IGV alignments and sashimi plots